requirements.txt
selenium
python-dotenv
pydantic
playwright
